requirements.txt
fastapi
uvicorn
tensorflow
python-multipart
Pillow
numpy